Refactored structure of the request object
